Switch benchmark example to RoBERTa (#13)
ResNet-50 currently takes longer to compile than RoBERTa does.  Swap out
the model to RoBERTa so the user spends less time waiting for model
compilation.  We'd switch the visualize example too, but RoBERTa is a
more complicated model that pushes the limits of Netron -- ResNet-50 is
more reasonably sized and works better with Netron, even though it takes
longer to compile in MAX.

Co-authored-by: Zheng Hao Tan <zheng@modular.com>