Update setuptools before installing dependencies from requirements.txt (#17)
It may throw an error on some OSes and architectures (Ubuntu 20 on
Graviton) if there's an old `setuptools` installed. So we upgrade it
before installing dependencies from `requirements.txt`.
The version range in `requirements.txt` is not enough because pip
installs dependencies from `requirements.txt` all at once using older
`setuptools`.
I'll keep the version range in `requirements.txt` as a "message" for
users.