Stable Diffusion: Consolidate model downloads into a top level, gitignored directory [1/N] (#52)
This is a first of `N` PRs moves a subset of downloaded example models
into a central, gitignored `models` subdirectory. By the end of it, we
won't have to re-download the same models across different examples /
scripts within this repository.

This PR covers `Stable Diffusion` examples within the `inference`
subdirectory.