RoBERTa: Consolidate common, downloaded models into a top level, gitignored directory [2/N] (#53)
This is 2 of `N` PRs moves a subset of downloaded example models into a
central, gitignored `models` subdirectory. By the end of it, we won't
have to re-download the same models across different examples / scripts
within this repository.

This PR covers `RoBERTa` examples within `inference/` and `tools/`. We
skip `notebooks/` for now as that is a standalone workflow that doesn't
rely on a prior downloaded model.