[MAX examples] Restore the performance showcase example, with TensorFlow removed (#38357)
In PR #38159, the performance showcase example was removed from the MAX
examples due to its reliance on TensorFlow models for the `run_max` part
of the benchmark. This replaces the TensorFlow models for the RoBERTa
and CLIP MAX benchmarks with TorchScript versions of either. It also
removes the TensorFlow benchmarks and other references to TensorFlow.

I've run these benchmarks on macOS and Ubuntu, and the results for the
TorchScript MAX benchmarks align with what they were for the previous
TensorFlow MAX runs.

MODULAR_ORIG_COMMIT_REV_ID: 8106a037d0218a4207c0108415d541d596f36905