[MAX][Examples] Avoid using `__mlir_op` in stable diffusion example (#38997)
[MAX][Examples] Avoid using `__mlir_op` in stable diffusion example

This was a suggested cleanup.
We prefer not to expose `__mlir_op`, it simply is not a nice api or to
read in general.
Turns out that it is pretty easy to build the DTypePointer via address
and avoid the `__mlir_op`.

MODULAR_ORIG_COMMIT_REV_ID: acc8c0ef525e0da773c5ae45c726152be235596a