[MAX][Examples] Yolo fix for remote linux (#39006)
On python-opencv only in linux there is a licence issue so the h264
encoder is not included with the prebuilt ffmpeg. This means you can't
view the generated video when on a remote linux machine, as vscode
doesn't support the included codecs.

This fixes that issue by using ffmpeg from ubuntu 22.04 which has the
h264 encoder.

It also opens the file automatically if running in vscode, only
downloads the file if not running in webcam mode, and cleans up the
args.

MODULAR_ORIG_COMMIT_REV_ID: 342d31219ab8963b1172d518278cba0ebfe5453a