Add finished code for **** inference tutorial
This is the finished code that users will create themselves when they
finish the "Run an **** model with Python" tutorial.

This code exists in the MAX repo primarily for CI testing, to ensure
that the tutorial code is functional, but also as a reference for users
to compare their finished work (or to cheat).

Notice that this creates a new top-level `tutorials` path in the MAX
repo, alongside the `examples` path. This is important because the
finished code for tutorials should not be the same as the other code
examples. For example, the code for this ResNet tutorial is much more
simple than the other "inference" code example using ResNet. This is
crucial for simplicity when building the project in a tutorial. The code
in each tutorial project is designed to work stand-alone and does not
depend on other code within the repo—it is fully self-contained in its
own folder within the `tutorials/` path.

MODULAR_ORIG_COMMIT_REV_ID: b82925dd2fb56542b74b65600a69986ba8e33242