[mojo-stdlib] Add a `List.swap_elements` method to address exclusivity issues.
It is very common to want to swap two elements of a List, and
`swap(a[i], a[j])`
is now (correctly) diagnosed as being an exclusivity problem because the
two
elements may alias. To solve this, we introduce a new method that takes
the
elements by index and handles it internally.

MODULAR_ORIG_COMMIT_REV_ID: 791e09ca01c36894f99cb66b24100014e17abe5d