[Pipelines] Make sure a default HF repository is used when loading local weights
We use the repository name to set up the AutoTokenizer for tokenization,
templating, etc. When loading local weights, that repository was not
being set, so it was causing serving to fail. This sets the default to
Modular's Hugging Face repository, although that can be overridden using
`--huggingface-weights`.

MODULAR_ORIG_COMMIT_REV_ID: 6ee466b087328a200927a38b5cef6d7bacb27513