[SDK] Don't mistake driver results for batch dim
Correctly marshal the batch logits out of the model execute in the
opaque path.
Previously this was mistakenly treating the `len(results)` of the
driver execute method as the batch dim.
This is a bit confusing because while the graph only returns a single
result: a batch of logits `(batch, seq_len, ntokens)`, the driver API
always returns a list.

In this case it's a single-element list, so just take the single element
"batch logits" out of the list.
After this, the shape of logits returned by the naive `_execute` and
opaque `_execute_opaque` paths match.

MODULAR_ORIG_COMMIT_REV_ID: fec6a61a33060f243afeffae5b3ce9d1fe5d71ca