[SDK] Set mask dtype to float32 for Q4_K
Set the mask dtype to float32 explicitly for `Q4_K`.
Previously we were casting to `params.dtype` in the graph.
This is to account for the fact that we build the attention mask outside
of the graph in NumPy, which doesn't have bfloat16.
However in doing so we also have to account for the fact that
`params.dtype` is uint8 for quantized dtypes.

So as a cleaner solution expose a `mask_dtype` method on the
`**************` class, which has special behaviour when
`QuantizationEncoding` is set.

MODULAR_ORIG_COMMIT_REV_ID: 34dcdaf875e6ca139049b87636ad9ad74ee95fab