[graph-api] Correctly slice out batch size in naive llama
Naive llama was treating the cache as having the same batch size as
tokens.
This is not the case. The cache can have a larger batch size.
As such, we need to slice out the batch size we care about.

This bug should be caught in the dynamic batching test.
It is not currently caught due to MSDK-1082.

Since this was found manually, I thought it would be best to just push a
patch.

MODULAR_ORIG_COMMIT_REV_ID: 7e8ad20c82082426048742d4ccf7f9de0b147672